requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.1
faiss-cpu==1.9.0
numpy>=1.26.0
rank-bm25==0.2.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
tiktoken>=0.7.0
httpx==0.25.1
jinja2==3.1.2
requests==2.31.0
